Move components into a separate folder @ app-module.
