+++ console/src/app/modules/app-module/app.module.ts
@@ -1,27 +1,21 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
-
-import { AppComponent } from './components/app-component/app.component';
-import { AppRoutingModule } from './app-routing.module';
+import { NgModule } from "@angular/core";
+import { AppComponent } from "./components/app-component/app.component";
+import { HeaderComponent } from "src/app/header/header.component";
+import { RunsTableComponent } from "./components/runs-table/runs-table.component";
+import { MongooseRunStatusIconComponent } from "./components/runs-table/mongoose-run-status-icon/mongoose-run-status-icon.component";
+import { RunsTableRootComponent } from "./components/runs-table/runs-table-root/runs-table-root.component";
+import { RunStatisticsComponent } from "./components/run-statistics/run-statistics.component";
+import { RunStatisticLogsComponent } from "./components/run-statistics/run-statistic-logs/run-statistic-logs.component";
+import { RunStatisticsChartsComponent } from "./components/run-statistics/run-statistics-charts/run-statistics-charts.component";
+import { DateFormatPipe } from "src/app/common/date-format-pipe";
+import { FormsModule, ReactiveFormsModule } from "@angular/forms";
+import { HttpClientModule } from "@angular/common/http";
+import { AppRoutingModule } from "./Routing/app-routing.module";
+import { MongooseSetUpModuleModule } from "../mongoose-set-up-module/mongoose-set-up-module.module";
+import { ControlApiService } from "src/app/core/services/control-api/control-api.service";
+import { MonitoringApiService } from "src/app/core/services/monitoring-api/monitoring-api.service";
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-
-import { MongooseRunStatusIconComponent } from './components/runs-table/mongoose-run-status-icon/mongoose-run-status-icon.component';
-
-// NOTE: NPM dependencies
-import { HeaderComponent } from '../../header/header.component';
-
-import { ControlApiService } from '../../core/services/control-api/control-api.service';
-import { MonitoringApiService } from '../../core/services/monitoring-api/monitoring-api.service';
-import { RunStatisticsComponent } from './components/run-statistics/run-statistics.component';
-import { RunStatisticLogsComponent } from './components/run-statistics/run-statistic-logs/run-statistic-logs.component';
-import { RunStatisticsChartsComponent } from './components/run-statistics/run-statistics-charts/run-statistics-charts.component';
-import { DateFormatPipe } from '../../common/date-format-pipe';
-import { RunsTableComponent } from './components/runs-table/runs-table.component';
-import { RunsTableRootComponent } from './components/runs-table/runs-table-root/runs-table-root.component';
-import { MongooseSetUpModuleModule } from '../mongoose-set-up-module/mongoose-set-up-module.module';
-
+import { BrowserModule } from '@angular/platform-browser';
 
 @NgModule({
   declarations: [
